+++ workspaces/ai-integrations/packages/backend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # backend
 
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [6af0114]
+- Updated dependencies [c2e3749]
+- Updated dependencies [56c40be]
+- Updated dependencies [67be6ce]
+  - @red-hat-developer-hub/backstage-plugin-catalog-backend-module-catalog-backend-module-ai-resource-extensions@0.2.0
+  - @red-hat-developer-hub/backstage-plugin-catalog-backend-module-model-catalog@0.10.0
+  - @red-hat-developer-hub/backstage-plugin-catalog-techdoc-url-reader-backend@0.6.0
+  - @red-hat-developer-hub/backstage-plugin-ai-experience-backend@0.4.2
+  - @red-hat-developer-hub/backstage-plugin-kserve-kubeflow-connector-backend@0.1.1
+
 ## 0.0.13
 
 ### Patch Changes